refactor(app): move SetupScreens to organisms
One fewer import from pages
