+++ app/src/organisms/ODD/ProtocolSetup/ProtocolSetupOffsets/index.tsx
@@ -15,7 +15,7 @@ import type { LabwareOffset } from '@opentrons/api-client'
 import { useToaster } from '../../../../organisms/ToasterOven'
 import { ODDBackButton } from '../../../../molecules/ODDBackButton'
 import { FloatingActionButton, SmallButton } from '../../../../atoms/buttons'
-import type { SetupScreens } from '../../../../pages/ODD/ProtocolSetup'
+import type { SetupScreens } from '../types'
 import { useMostRecentCompletedAnalysis } from '../../../LabwarePositionCheck/useMostRecentCompletedAnalysis'
 import { TerseOffsetTable } from '../../../../organisms/LabwarePositionCheck/ResultsSummary'
 import { getLabwareDefinitionsFromCommands } from '../../../../molecules/Command/utils/getLabwareDefinitionsFromCommands'
