+++ app/src/organisms/ODD/ProtocolSetup/types.ts
@@ -1,0 +1,9 @@
+export type SetupScreens =
+    | 'prepare to run'
+    | 'instruments'
+    | 'modules'
+    | 'offsets'
+    | 'labware'
+    | 'liquids'
+    | 'deck configuration'
+    | 'view only parameters'
